refactor: extract mcp protocol framing
